requirements.txt
sympy
ply
pyyaml
z3-solver
